Prevent weak type promotion to reduce recompilation (#1963)
Resolves #1581 

Some of the attributes in optimizable classes are cast to float by
`float()` function which returns a Python value, but internally JAX
converts them to `jnp.float64` before jit, this causes recompilation in
some cases. Technically, the number of compilations shouldn't be more
than 2, but this can be the cause of some recompilations.

I know there has been some discussion on not using `jnp.float64` before
to leave some room for #1048, but I don't see that working in the near
term.

```python
jax.config.update('jax_explain_cache_misses', True)
```
This was very useful to pinpoint the exact cause of recompilation in
#1581

---------

+++ desc/optimize/least_squares.py
@@ -285,7 +285,7 @@ def lsqtr(  # noqa: C901
     if g_norm < gtol:
         success, message = True, STATUS_MESSAGES["gtol"]
 
-    alpha = 0.0  # "Levenberg-Marquardt" parameter
+    alpha = jnp.float64(0.0)  # "Levenberg-Marquardt" parameter
 
     while iteration < maxiter and success is None:
 
@@ -314,7 +314,7 @@ def lsqtr(  # noqa: C901
         actual_reduction = -1
 
         # theta controls step back step ratio from the bounds.
-        theta = max(0.995, 1 - g_norm)
+        theta = jnp.float64(max(0.995, 1 - g_norm))
 
         while actual_reduction <= 0 and nfev <= max_nfev:
             # Solve the sub-problem.
